✅ Sqrt(x) | Easy | Java | Runtime: 56 ms (5.64%) | Memory: 42.9 MB | 2026-07-19

+++ leetcode/easy/0069-sqrtx/solution.java
@@ -3,9 +3,9 @@ class Solution {
         if (x <= 1)
             return x;
 
-        for (int i = 1; i <=x / 2; i++) {
+        for (long i = 1; i <=x / 2; i++) {
             if (i * i > x)
-                return i - 1;
+                return (int)(i-1);
 
         }
         return x / 2;
